Refactor navigation and variable names
Split nav.html into two different files - one for
signed in user and one for guest. Depending on the user, they
will get a different navigation bar.

Fix session variable names to be clearer in how what they stand for

Add protection to signup/signout/signin pages so they can't be accessed
at all times. The user needs to be signedout to access the signout page
and needs to be signed out to have access to the signin/signup pages.